+++ src/main/Args.ts
@@ -1,5 +1,5 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { version } = require("./../package.json");
+ 
+import { version } from "./../package.json";
 
 export interface AppArgs {
   figmaUrl: string;