+++ src/renderer/Panel/store/tabs.svelte.ts
@@ -14,7 +14,7 @@ function addTab(data: Types.AddTabProps) {
       focused: data.focused,
       isUsingMicrophone: false,
       isInVoiceCall: false,
-      loading: true,
+      loading: data.loading ?? true,
     },
   ].sort((a, b) => (a.order > b.order ? 1 : -1));
 }