+++ src/utils/Main/fs.ts
@@ -3,8 +3,8 @@ import * as fs from "fs";
 export async function mkdirIfNotExists(path: string) {
   try {
     await fs.promises.mkdir(path);
-  } catch (error: any) {
-    if (error?.code !== "EEXIST") throw error;
+  } catch (error) {
+    if ((error as NodeJS.ErrnoException)?.code !== "EEXIST") throw error;
   }
 }
 
@@ -12,7 +12,7 @@ export async function access(path: string): Promise<boolean> {
   try {
     await fs.promises.access(path);
     return true;
-  } catch (error) {
+  } catch {
     return false;
   }
 }
@@ -22,7 +22,7 @@ export function accessSync(path: string): boolean {
     fs.accessSync(path);
 
     return true;
-  } catch (error) {
+  } catch {
     return false;
   }
 }